cmake: validate `CURL_DEFAULT_SSL_BACKEND` config value
Before this patch CMake builds accepted any value and it was used at
runtime as-is. This patch make sure that the selected default backend
is also enabled in the build. It also enforces a full lowercase value.

This improves reproducibility and brings CMake in sync with autotools
which already worked like described above.

Follow-up to 0368a556f0fd5e17951f4fe45314b99f29bd3f55 #11774

Closes #11998